+++ tooling/xtask/src/tasks/workflows.rs
@@ -8,6 +8,7 @@ mod nix_build;
 mod release_nightly;
 mod run_bundling;
 
+mod run_tests;
 mod runners;
 mod steps;
 mod vars;
@@ -23,7 +24,7 @@ pub fn run_workflows(_: GenerateWorkflowArgs) -> Result<()> {
         ("nix_build.yml", nix_build::nix_build()),
         ("run_bundling.yml", run_bundling::run_bundling()),
         ("release_nightly.yml", release_nightly::release_nightly()),
-        // ("run_tests.yml", run_tests::run_tests()),
+        ("run_tests.yml", run_tests::run_tests()),
         // ("release.yml", release::release()),
     ];
     fs::create_dir_all(dir)
