+++ tooling/xtask/src/tasks/workflows/release_nightly.rs
@@ -1,5 +1,6 @@
 use crate::tasks::workflows::{
     nix_build::build_nix,
+    run_tests,
     runners::{Arch, Platform},
     steps::NamedJob,
     vars::{mac_bundle_envs, windows_bundle_envs},
@@ -30,9 +31,9 @@ pub fn release_nightly() -> Workflow {
     .map(|(key, value)| (key.into(), value.into()))
     .collect();
 
-    let style = check_style();
-    let tests = run_tests(Platform::Mac);
-    let windows_tests = run_tests(Platform::Windows);
+    let style = run_tests::check_style();
+    let tests = run_tests::run_platform_tests(Platform::Mac);
+    let windows_tests = run_tests::run_platform_tests(Platform::Windows);
     let bundle_mac = bundle_mac_nightly(&[&style, &tests]);
     let linux_x86 = bundle_linux_nightly(Arch::X86_64, &[&style, &tests]);
     let linux_arm = bundle_linux_nightly(Arch::ARM64, &[&style, &tests]);
@@ -80,56 +81,9 @@ pub fn release_nightly() -> Workflow {
         .add_job(update_nightly_tag.name, update_nightly_tag.job)
 }
 
-fn check_style() -> NamedJob {
-    let job = release_job(&[])
-        .runs_on(runners::MAC_DEFAULT)
-        .add_step(
-            steps::checkout_repo()
-                .add_with(("clean", false))
-                .add_with(("fetch-depth", 0)),
-        )
-        .add_step(steps::cargo_fmt())
-        .add_step(steps::script("./script/clippy"));
-
-    named::job(job)
-}
-
-fn release_job(deps: &[&NamedJob]) -> Job {
-    let job = Job::default()
-        .cond(Expression::new(
-            "github.repository_owner == 'zed-industries'",
-        ))
-        .timeout_minutes(60u32);
-    if deps.len() > 0 {
-        job.needs(deps.iter().map(|j| j.name.clone()).collect::<Vec<_>>())
-    } else {
-        job
-    }
-}
-
-fn run_tests(platform: Platform) -> NamedJob {
-    let runner = match platform {
-        Platform::Windows => runners::WINDOWS_DEFAULT,
-        Platform::Linux => runners::LINUX_DEFAULT,
-        Platform::Mac => runners::MAC_DEFAULT,
-    };
-    NamedJob {
-        name: format!("run_tests_{platform}"),
-        job: release_job(&[])
-            .runs_on(runner)
-            .add_step(steps::checkout_repo())
-            .add_step(steps::setup_cargo_config(platform))
-            .add_step(steps::setup_node())
-            .add_step(steps::cargo_install_nextest(platform))
-            .add_step(steps::clear_target_dir_if_large(platform))
-            .add_step(steps::cargo_nextest(platform))
-            .add_step(steps::cleanup_cargo_config(platform)),
-    }
-}
-
 fn bundle_mac_nightly(deps: &[&NamedJob]) -> NamedJob {
     let platform = Platform::Mac;
-    let job = release_job(deps)
+    let job = steps::release_job(deps)
         .runs_on(runners::MAC_DEFAULT)
         .envs(mac_bundle_envs())
         .add_step(steps::checkout_repo())
@@ -144,7 +98,7 @@ fn bundle_mac_nightly(deps: &[&NamedJob]) -> NamedJob {
 
 fn bundle_linux_nightly(arch: Arch, deps: &[&NamedJob]) -> NamedJob {
     let platform = Platform::Linux;
-    let mut job = release_job(deps)
+    let mut job = steps::release_job(deps)
         .runs_on(arch.linux_bundler())
         .add_step(steps::checkout_repo())
         .add_step(steps::setup_sentry())
@@ -170,7 +124,7 @@ fn bundle_windows_nightly(arch: Arch, deps: &[&NamedJob]) -> NamedJob {
     let platform = Platform::Windows;
     NamedJob {
         name: format!("bundle_windows_nightly_{arch}"),
-        job: release_job(deps)
+        job: steps::release_job(deps)
             .runs_on(runners::WINDOWS_DEFAULT)
             .envs(windows_bundle_envs())
             .add_step(steps::checkout_repo())
@@ -184,7 +138,7 @@ fn bundle_windows_nightly(arch: Arch, deps: &[&NamedJob]) -> NamedJob {
 fn update_nightly_tag_job(deps: &[&NamedJob]) -> NamedJob {
     NamedJob {
         name: "update_nightly_tag".to_owned(),
-        job: release_job(deps)
+        job: steps::release_job(deps)
             .runs_on(runners::LINUX_CHEAP)
             .add_step(steps::checkout_repo().add_with(("fetch-depth", 0)))
             .add_step(update_nightly_tag())
