+++ tooling/xtask/src/tasks/workflows/steps.rs
@@ -118,6 +118,19 @@ pub(crate) struct NamedJob {
     pub job: Job,
 }
 
+pub(crate) fn release_job(deps: &[&NamedJob]) -> Job {
+    let job = Job::default()
+        .cond(Expression::new(
+            "github.repository_owner == 'zed-industries'",
+        ))
+        .timeout_minutes(60u32);
+    if deps.len() > 0 {
+        job.needs(deps.iter().map(|j| j.name.clone()).collect::<Vec<_>>())
+    } else {
+        job
+    }
+}
+
 // (janky) helper to generate steps with a name that corresponds
 // to the name of the calling function.
 pub(crate) mod named {
